move selector component to a separate file
